pyproject.toml
[project]
name = "crawler-rag-dochelper"
version = "1.0.0"
description = "A RAG system with hybrid search that crawls documentation and provides intelligent Q&A"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
authors = [
    { name = "NFRohan" }
]
keywords = ["rag", "langchain", "pinecone", "hybrid-search", "bm25", "documentation"]

dependencies = [
    "beautifulsoup4",
    "tiktoken",
    "openai",
    "unstructured",
    "nltk",
    "fastapi",
    "jinja2",
    "uvicorn",
    "streamlit",
    "streamlit-chat",
    "tqdm",
    "langchainhub",
    "firecrawl-py",
    "langchain-community", # Updated to align with core 1.0 ecosystem
    "langsmith",
    "pillow", # Required for Python 3.13
    "langchain-tavily",
    "langchain-pinecone",
    "certifi",
    "langchain-chroma",
    "langchain-cohere",
    "cohere",
    "python-dotenv",
    "langchain-huggingface",
    "langchain-google-genai",
    "sentence-transformers",
    "langchain", # Bumped to match current stable release
    "langchain-text-splitters",
    "pinecone", # Official Pinecone SDK (renamed from pinecone-client)
    "pinecone-text",
    "mmh3", # Fixes Windows build error
    "numpy", # Required for Python 3.13 wheels
    "greenlet",
    "aiohttp",
    "httpx", # Async HTTP client for fetching llms.txt
    "tavily-python", # Tavily SDK for content extraction
    "tavily",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "ipykernel",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 100

